17/07/2015

+++ worker/contribute.php
@@ -20,6 +20,33 @@ if(!$query){
 	exit('Database error : '.$err);
 }
 
+//check if the target number of contributions is reached
+$query = $db->query("SELECT target FROM question WHERE id=$question");
+if(!$query){
+	$err = $db->error;
+	$db->close();
+	exit('Database error : '.$err);
+}
+$target = $query->fetch_assoc()['target'];
+//get the actual number of contributions
+$query = $db->query("SELECT count(*) FROM contribution WHERE id_question=$question");
+if(!$query){
+	$err = $db->error;
+	$db->close();
+	exit('Database error : '.$err);
+}
+$count = $query->fetch_assoc()['count(*)'];
+
+if($count>=$target){
+	$query = $db->query("UPDATE question SET status='completed' WHERE id=$question");
+	if(!$query){
+		$err = $db->error;
+		$db->close();
+		exit('Database error : '.$err);
+	}
+}
+
+
 //remove the attribution if existing
 $db->query("DELETE FROM attribution WHERE id_question=$question AND id_user=$userid");
 if(!$query){
@@ -28,5 +55,6 @@ if(!$query){
 	exit('Database error : '.$err);
 }
 
+$db->close();
 ?>
 <p>Contribution successfully registered.</p>
